Implement snyk plugin/task properties
Description
===========

Initial task/extension properties setup for snyk test and
snyk monitor command.
All currently available cli arguments are setup as task properties
which are mappend to the plugin extension which in turn get
the default values from the environment/gradle properties. This
allows a maximum of flexibility for setting snyk up on cli system.

Changes
=======

* ![ADD] snyk plugin/task properties