Add snyk download task (#3)
## Description

Add an auto-download task for Snyk. It needs to know the version it needs to download, and where to download the file. These can be set through extension or through the task itself, defaulting to version `v1.840.0` (latest by the time of coding), the download path defaults to `GRADLE_HOME/atlas-snyk`, and the executable name defaults to "snyk", changing to "snyk.exe" on windows. 

The file is downloaded to a folder matching its version using groliphant wrapper of gradle distribution downloader. For instance, a version tagged "v.1.19", will be downloaded to `$downloadPath/v1.19/$executableName`.

PS: I believe that it would be way better for us to provide the snyk file itself to the SnykTask, instead of a path divided into directory and name. I haven't changed it though, because I don't know the exact reasons of things being the way they are right now.

## Changes

* ![ADD] snyk download logic