Add snyk runtime strategy implementation (#6)
Description
===========

This patch adds a few more properties to the snyk extension
to configure how snyk will hook itself into the build lifecycle.
For most plugins there is only one reasonable way of adding tasks
to the task collection. With snyk multiple viable ways are possible.
So here we define a strategy list to configure how the configured
snyk tasks are added to the build lifecycle.

To configure this one needs to set the `strategies` property on
the snyk extension. It is of type `List<String>` to make it foreward
compatible and help ease the API.

If the list is empty, which is the default, then the snyk plugin
won't try to inject itself into the build.
There are three predefined strategy values

| Name            | Description                            |
| --------------- | -------------------------------------- |
| monitor_check   | Run snyk monitor when executing checks |
| monitor_publish | Run snyk monitor when publishing       |
| test_check      | Run snyk test when executing checks    |

Setting all three values would mean that `snyk monitor` and
`snyk test` would run when executing checks. Or only `snyk monitor`
during publish when tests are excluded. This will hold true
for any task of type `Monitor` and `Test`.

Additionally one can also configure the `check` task name and the
`publish` task name. The default values are: `check`, `publish`.

Changes
=======

* ![ADD] snyk runtime strategy implementation