Improve task test setup
Description
===========

While working on an addition I saw that we test the task implementation
with the plugin applied during test. This is generally not a problem but
brings some testing issues with it. The assumption of the task integration
tests is to test the task behavior under isolation and with it's defaults.

What happened was that some tests fail when the plugin is not setting up
the correct conventions. In my case it lead to a false positive for some
property tests because the case only worked as expected when the plugin
configures some conventions. I understand that this will be 98% of the time
the case for us but we then cannot say that the task type under test
is actually well isolated from the plugin itself to be used elsewhere.
We have to remember that the plugin is for us just a huge convention mapper.
And we could easily provide multiple different versions with different
opinioted settings. Therefore testing that the task classes work without
the plugin applied is a must.

This patch is doing just that. I removed the line to apply the plugin
in the `SnykTaskIntegrationSpec`. All other changes are fixes to
get the tests working again.

As an example. I had to make sure that we set the fake `snyk` client
on the task itself not the extension.

For the real `integration` test I added a small custom spock extension
to download snyk and provide a file path.

Since this patch derived from a different task I had to keep some prepartion
changes. I split the `TestIntegrationSpec` further down into yet another base
class. This will make sense with the next patch.

Changes
=======

* ![IMPROVE] task test setup