Add missing convention mappings and tests
Description
===========

This patch adds some missing convention setups.
* autoDownload
* autoUpdate
* debug

The `debug` property had already a convention as it was bound to the
gradle log level. I decided to add basic override in case someone wants
to set debug also in the quiet or info level.

While writing the tests I found a few cases missing. Most noteably the
missing tests for the Report class.

I also added a way of clearing the test environment from all declared environment
variables which this plugin defines. This means that global values on a development
or ci machine should not dictate the outcome of tests.

Changes
=======

* ![ADD] convention setup for `autoDownload` property
* ![ADD] convention setup for `autoUpdate` property
* ![ADD] convention setup for `debug` property
* ![IMPROVE] tests cases
* ![IMPROVE] test environment setup