Improve latest version support
Description
===========

I want to switch to the model of always using `latest` as the default version.
This has two issues.

* we might break in the future
* latest needs to be treated differently for the up-to-date check

This patch addresses the second issue since the first one can't really
be addressed. The fact is that `snyk` creates release very often
and it would be helpfull if a `latest` can just be the default and
the global `SNYK_VERSION` variable or gradle property (`snyk.version`) as
a way of fixing compatibility issues as a workaround.

To address the `up-to-date` check I added a new inbetween provider called
`versionToInstall` which is also tracked as an `Input` property.
It maps over the provided `version` and checks if the value is `latest`.
If it is latest it will fetch the actual latest version from snyk release.json
and returns this instead. This only makes sure that the `up-to-date` check will
realize that `latest` is out of date when `snyk` released a new version.

I decided to remove the property `autoUpdate` as adding this into the mix
creaated quite the complicated mess. And to be honest I don't see a scenario
where we want to use `latest` but keep the installed version.

To test this I went all out and added [mock-server] library.
This allowed me to override and forward calls to the static snyk endpoint.
To do this I changed some url constants to fetch the hostname and port from
system settings. I think this is quite clean allows to test this case quite easily.
The testcase using it is:

run the task which calls `latest` release.json and fetches an old version.
(this call is forwared to an old release json)

Then clear the forwards and call the task again and make sure it is not up
to date and that the version pulled is different.

Changes
=======

* ![IMPROVE] latest version support

[mock-server]: https://www.mock-server.com