Improve registered subproject names based on projectName property when set
Description
===========

The first implementation for registered sub projects or project files
took the gradle root project name and appended the subproject path / filename.
If one overrides the projectName property in the root project means
All subprojects would need to be manually configured.

I adjusted the logic here by using a `Provider` instead of a `String`
type and evaluate the property lazy. Means that the registered
`projectName` properties will be based on the root `projectName` property
when set or fall back to the gradle project name.

Changes
=======

* ![IMPROVE] registered subproject names based on root `projectName` property when set