Add SnykToHtmlInstall task
Description
===========

This patch adds a new installer task for the `snyk-to-html` utility.
The tool can be used to convert `json` reports generated by `snyk test`
to `html` report pages. This is a preparation patch to add an installer
task. The task is not yet wired in the plugin. I mainly added the
functionality to download the tool. The basic setup on snyk side
is very similar to the cli tool. There is a slight exception that
this tool won't be published with a `releases.json` file. But
the whole download setup is so generic that I created an abstract
base class which the new install task and the `SnykInstall` task both
extend from. I did the same for the integration tests.

Changes
=======

* ![ADD] `SnykToHtmlInstall` task