Fix reporting task property mapping in registered project
Description
===========

During implemetation of a new feature I found that the properties
from the registered extension for a sub project or package file
wasn't mapped to the reports task of the registered project.
Easy enough to fix since the mapping was missing.

I added now the missing tests to test all property mappings from the
generated tasks for the registered sub projects etc. I left this out
during implementation of the feature to not loop over the list yet again.
As it turns out this was a failure.

Changes
=======

* ![FIX] reporting task property mapping in registered project