Fix reporting conventions
Description
===========

The last patch still didn't fix my issue even though now
all tests for the convention mapping had been added. As it turns out
the tests only run without the plugin conventions. So the task defaults
work under test and isolated but the plugin convention breaks it.

During my investigations why this happens I saw 3 bigger mishaps:

* I introduced two compeding ways to set the json-output path argument
* one should not use `reports.json.enabled` and `reports.json.required` in mixed mode
* the shared traids for task and extention classes contained too much code for each context.

So I first set out and straightened the traids. This would not have been technically
neccesary but I think it is cleaner that way. There is still a lot of improvment when
it comes to naming and how many layers we need.

What I did is to split the task relevant APIs from the extension APIs.
Means that all the `OptionMapper` implementations are split into seperate traits only
needed for the task implementations. The extension has no need for these.

The main issue that I needed to address was the way we configure the
`sarifOutputPath` / `jsonOutputPath` property. This property directly controls
the matching flag for the snyk cli. It needs to be setable for the task instance but
I removed all mappings to the extension and the convention providers. Here we use
the reports dir and report enable properties. This is cleaner and makes a lot more sense.

While adjusting the tests I ran into the issue of using `reports.json.enabled` instead of
`reports.json.required` (the new provider api variant). This can lead to unwaned sideeffects
so I made sure that we only use one of the APIs.

Changes
=======

* ![FIX] reporting output path conventions