Fix snyk token handling
Description
===========

During testing I saw that snyk monitor only works by coincidence since
we ask for the same environment property name `SNYK_TOKEN` as a convention
to set the `token` property. The token property is actually not needed
as a common or base property since only commands whi interact with snyk backend
need a token to authenticate. I left it in `SnykTaskSpec` for the moment since
we lack a generic way to describe custom environment variables from properties
similar to commandline arguments. A feature request is filed for [wooga/gradle-commons#14].

I did a very minimalistic implemenation to make sure that the token value is passed
as environment variable to `snyk` cli.

Changes
=======

* ![FIX] `snyk` token handling

[wooga/gradle-commons#14]: https://github.com/wooga/gradle-commons/issues/14