Make snyk tasks failable with additional property (#45)
Description
===========

This patch adds a new helper property `ignoreExitValue` to
the base `SnykTask` class. This allows to make a single or all
snyk tasks failable. A usecase I see is to leave the configuration
alone (the default is the old default: fail on error, exception is `Report` task).

We can then set a global environment variable `SNYK_IGNORE_EXIT_VALUE=1` which
will activate the flag for all tasks.

Changes
=======

* ![IMPROVE] make `snyk` tasks failable with additional property `ignoreExitValue`